finzo-backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
groq==0.9.0
gunicorn==22.0.0
openpyxl==3.1.2
python-dotenv==1.0.1
